Extract target blank assertions into separate file
