+++ spec/e2e/accessibility_spec.js
@@ -1,25 +1,7 @@
 import { AxePuppeteer } from '@axe-core/puppeteer';
 import { toHaveNoViolations } from 'jest-axe';
 import { page, goto } from './support/browser';
-
-/**
- * @typedef SimplifiedLink
- *
- * @prop {string=} innerText
- * @prop {string=} href
- * @prop {string=} className
- * @prop {string=} target
- */
-
-/**
- * @param {SimplifiedLink} a
- */
-function toNotHaveTargetBlank(a) {
-  return {
-    pass: a.target != '_blank',
-    message: () => `Link "${a.innerText}" to ${a.href} had target=_blank, but it should not have`,
-  };
-}
+import { getLinks, toNotHaveTargetBlank } from './support/target-blank';
 
 expect.extend(toHaveNoViolations);
 expect.extend({ toNotHaveTargetBlank });
@@ -50,19 +32,8 @@ describe('accessibility', () => {
         .analyze();
       expect(results).toHaveNoViolations();
 
-      /** @type {SimplifiedLink[]} */
-      let links = await page.evaluate(() =>
-        Array.prototype.map.call(document.querySelectorAll('a'), (a) => {
-          // There's a weird serialization boundary here so we encode the links as a structure
-          return {
-            innerText: a.innerText.trim(),
-            href: a.href,
-            className: a.className,
-            target: a.target,
-          };
-        }),
-      );
-      Array.from(links).forEach((a) => {
+      const links = await getLinks(page);
+      links.forEach((a) => {
         expect(a).toNotHaveTargetBlank();
       });
     },
