+++ spec/e2e/support/target-blank.js
@@ -1,0 +1,36 @@
+/**
+ * @typedef SimplifiedLink
+ *
+ * @prop {string=} innerText
+ * @prop {string=} href
+ * @prop {string=} className
+ * @prop {string=} target
+ */
+
+/** @type {import('puppeteer').Page} */
+/** @return {Promise<SimplifiedLink[]>} */
+async function getLinks(page) {
+  return await page.evaluate(() =>
+    Array.prototype.map.call(document.querySelectorAll('a'), (a) => {
+      // There's a weird serialization boundary here so we encode the links as a structure
+      return {
+        innerText: a.innerText.trim(),
+        href: a.href,
+        className: a.className,
+        target: a.target,
+      };
+    }),
+  );
+}
+
+/**
+ * @param {SimplifiedLink} a
+ */
+function toNotHaveTargetBlank(a) {
+  return {
+    pass: a.target != '_blank',
+    message: () => `Link "${a.innerText}" to ${a.href} had target=_blank, but it should not have`,
+  };
+}
+
+export { getLinks, toNotHaveTargetBlank };
